-renamed 'xfPending' to 'bfPending' -changed style rules to be less 'intrusive'.
Controls + Repeats get the 'bfPending' class when there are calls to the server that not have returned yet. This can be used to visually notify the user that something is still going on. The styles should be handled with care and must be tested for every concrete theme developed for a site.