Fix sporadic travis test failures
Some ContainerTest unit tests could fail depending on the order the tests were executed. The problem is that Xerces maintains a global table with counts of registered listeners in org.apache.xerces.dom.LCount, so if a test method that does not properly unregister its listeners, it may influence the behavior of subsequently executed tests.

The fix unregisters all test listeners at the end of each test and adds missing event listeners where they are necessary for the expected behavior.
Conflicts:
	core/src/test/java/de/betterform/xml/xforms/ContainerTest.java