FIX: removed 10% chance of parseISODate() failing when parsing xs:dateTime values
xs:dateTime as implemented by Saxon does not pad the
fractional seconds with zeroes, leading to a parsing error
in ExtensionFunctionsHelper.parseISODate whenever the
time to be parsed contains a number of milliseconds
that is divisible by 10. As a consequence, the unit test XFormsExtensionFunctionsTest.testNow() failed in approx.
1 out of 10 executions.

This patch fixes the parsing function and tests the
modified behavior.