More robust error logging in ErrorServlet
There were several cases where internal processing in the ErrorServlet would throw an exception, resulting in the overly broad (and often wrong) message „XFormsException: Prefix xf has not been declared“ in the log.

This patch logs whatever information has been collected up to the point of the exception, resulting in much more helpful messages. It also uses LOGGER to output the error document instead of dumping it to System.out .